% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrpicker.R
\name{updateDateTimeRangePickerInput}
\alias{updateDateTimeRangePickerInput}
\title{Update a datetime range picker input}
\usage{
updateDateTimeRangePickerInput(session, inputId, from, to)
}
\arguments{
\item{session}{the Shiny \code{session} object}

\item{inputId}{the id of the datetime range picker widget to be updated}

\item{from, to}{new values for the datetime range picker widget}
}
\value{
No returned value, called for side effect.
}
\description{
Change the values of a datetime range picker input.
}
